[pylint] Use --output-format=json
Using `--msg-template` and a regex to parse the Pylint output didn't
properly handle some messages, e.g. the final "]" in "Dangerous default
value [] as argument" would be greedily interpreted as the end of the
"category/symbol" section of the line, resulting in malformed comments
such as this one:
https://fuchsia-review.googlesource.com/c/infra/recipes/+/615990/1..6/recipe_modules/fxt/api.py#b73

By switching to a JSON output format we get a much more easily parseable
output that is more resilient to varied outputs. Here's the JSON of the
comment that the pylint_parser will now emit for the same pylint result
mentioned above, based on running the updated version locally:

  {
    "category": "Pylint/warning/dangerous-default-value",
    "message": "Dangerous default value [] as argument.\nTo disable, add: # pylint: disable=dangerous-default-value",
    "path": "recipe_modules/fxt/api.py",
    "startLine": 17,
    "startChar": 4
  }

Bug: 1275987
Change-Id: I2807045426a33b2e91dd7732e8fd70bfa37b3d71
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3328042
Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#45965}