[pylint] Inspect pylint exit code
And fail the pylint_parser if the exit code is not one of the four exit
codes that are known to be non-fatal. This will make it easier to debug
when pylint fails and doesn't emit valid JSON output, because the
emitted error will correspond to the actual error rather than a JSON
parsing error.

Bug: 1275987
Change-Id: I954360dee73e9985b9ac5a076d7c08edeb176b0d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3328043
Commit-Queue: Oliver Newman <olivernewman@google.com>
Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#45966}