[Weetbix] Remove dependency on CV proto in ingestion code.
A small number of runs may not be ingested for variant analysis or
incorrectly ingested (i.e. presubmit runs ingested as non-presubmit
run) as a result of this change, as we transition from the old
to the new ingestion proto.

This was not mitigated as it is believed the impact is low
compared to the effort of landing the CLs in two parts.

This change also fixes the format of LUCI CV IDs used in ingestion,
which were previously of the form 'project/{LUCI_PROJECT}/runs/{RUN}'
and will now be of the form '{LUCI_PROJECT}/{RUN}'. The old
implementation did not align with the design intent as captured by
the the field description and makes it harder to link to the LUCI CV
page, e.g.

https://luci-change-verifier.appspot.com/ui/run/infra/8983502628854-1-b48282245a9f1d4f

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: I429755d86b6fbf01f612dc84e9aac94e99b02e02
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3340135
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#46044}