Modify chromium bootstrapper to participate in process termination.
Currently, the bootstrapper doe not do anything to handle the process
termination, so the bootstrapped executable never gets soft-terminated
to allow it to respond to the event. This changes it so that the
bootstrapper receives the signal and relies on it being automatically
propagated to the bootstrapped process.

led run showing cancellation of bootstrapped build:
https://luci-milo.appspot.com/raw/build/logs.chromium.org/chromium/led/gbeaty_google.com/886c64d9bdf620112abd63af83591964c1331639856f9f21b5e6f3a81f9e4f1c/+/build.proto

Change-Id: I7612ae911dd6b149eb6fbbc8d6e16156f6efa2f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3339952
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#46057}