[Weetbix] Configure Weetbix for 90 day data retention.
Data will be retained for 90 days, except for some ingestion-related
data:
- Chunks uploaded to GCS, for which PartitionTime is not available.
- Ingestion Control records, for which PartitionTime is not available.

This data will be retained for a longer period of 100 days from either
creation or last update (whichever is available), to ensure it is not
deleted before the corresponding ingested chunks are deleted.

Ingestion will now also verify ingested data is within an allowed
time range of -90 days to +1 day to ensure data is not deleted
immediately after creation (or is created so far in the future that
it escapes deletion). The window is generous to allow backfilling
of data and clock drift.

A follow up CL will apply the data retention policy to GCS.

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: I31e57eb2298a38887044e1c512f8aab47524df45
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3340146
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#46068}