[recovery]: Fixed a problem with executing servod commands.
This commit is a part of the PARIS project.

The conditions that checks that ServoHost was correctly determined
before executing any servod commands was inverted. This was causing
all attempts to execute the commands to terminate with message "servo
not found". This commit fixes this problem.

Bug=b:198809511
TEST=build, manual test, unittest

Change-Id: I2696469977a719920deed86d4d9c7bd51ce99b8d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3353167
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#46088}