[3pp] Fix refs for run_remotely
If the second-to-last digit of a CL number is '0' then
run_remotely_extract_repo_ref.py wouldn't return a valid ref for a CL.
For example, CL 11102 PS 3 should be 'refs/changes/02/11102/3' but we
were getting 'refs/changes/2/11102/3'. Fix that by using '%02d' in the
format string.

Change-Id: Iefb9a48f6f59241f687ac2845c19395db80fe589
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3359955
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Rob Mohr <mohrr@google.com>
Cr-Commit-Position: refs/heads/main@{#46106}