[Weetbix] Group variants of the same gtest and blink web test together.
This change defines a framework of rules which can be used to
control how test names are clustered. This framework is initially
used to cluster variants of the same google test and blink web test
together.

In follow-up change, the rules will be made configurable on a
per-LUCI project basis.

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: I22b2adcc569bafa27298e4897e541734d895f519
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3358658
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#46113}