[crosskylabadmin] Add UFS client
We need to add a UFS client to CrOSSkylabAdmin in order to determine the
pools associated with each labstation, which is needed for labstation
scheduling.

In this CL, we add some basic functionality for listing swarming
dimensions and instantiate-but-do-not-use the UFS client inside the
repair job.

BUG=b:206044795
TEST=make test

Change-Id: I4f5e9e1cb39c4be6ae6c9235061eaf85c13b760b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3322092
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#46141}