Add retries for downloading a file from gitiles.
In some instances when a build has a gitiles commit for a valid
revision, the bootstrapper will fail to be able to download the file
with a not found result. We expect that this is due to replication lag
for gitiles. This change adds some retries to prevent infra-failing in
these situations. Tweaking of the specific values will probably be
necessary in the future.

Bug: 1279551
Change-Id: I3c7796cbba77294ca24ab36d9a82bbd18041db5b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3366918
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#46152}