[Weetbix] Version project configuration.
Whenever project configuration is updated, set a new LastUpdated
time. This version should always be increasing for newer versions.

A timestamp was used in preference to an incrementing integer as
temporary deletion of a project's configuration would result in
the integer resetting to its starting value (e.g. 1) whereas a
timestamp can still be expected to increment when the configuration
is restored.

As configuration is only re-queried from LUCI Config every minute,
and config changes themselves are likely to be far more rare,
it is unlikely the LastUpdated time would ever go backwards due
to clock synchronisation issues. Nonetheless, a check is introduced
to verify this (for the case where config is not deleted and
re-created.)

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: I48a02d671a19e4d60f0c1b7b54e015917684141b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3358673
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#46170}