[labpack] Describe environment labpack is run in.
As part of b:193934420, log collection for Paris, we need a better idea
of where the different kinds of logs of located. In order to get
trustworthy information, labpack should record information about where
it's located and what it sees there during teardown.

This should be run as the very last step.

BUG=b:193934420
TEST=ran labpack with patch described in b:193934420 comment #14 locally and observed result

Change-Id: I176b712d955841efb8535a012b1dcc3db25d3919
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3370586
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#46186}