[recovery]: Validate the battery using servod controls and update state.
This commit is a part of the PARIS project.

This commit adds the logic to verify that that the battery on DUT can
be charged. This is done using servod controls. Accordingly the
hardware state on the DUT is updated to reflect the result of check.

Bug=b:198809511
TEST=build, manual test

Change-Id: Iff1185934f3621e5790f595015c48954519d99b8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3364868
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#46195}