[Weetbix] Define cache for compiled project config.
Provide an accessor for a version of project config for which
test name clustering rules have been compiled. To avoid repetitive
regexp recompilation on every config access, the compiled form of
the config is cached.

This change also changes the behaviour of
config.SetTestProjectConfig() so that the config LastUpdated time
set via this method is returned by config.Project(...). This is
the more useful behaviour when writing integration tests.

BUG=1243174
TEST=INTEGRATION_TESTS=1 go test infra/appengine/weetbix/...

Change-Id: I0e7963d93ae2f7c7231ec9530f80a25606c744c6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3367520
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#46229}