[Weetbix] Add ConfigVersion field to ClusteringState table.
Define a field to track the version of configuration that was
used to cluster test results. Initially this is always populated
to the primordial configuration version, but once clustering is
updated to use configuration, this will be updated to the version
of configuration that was actually used.

BUG=1243174
TEST=INTEGRATION_TESTS=1 go test infra/appengine/weetbix/...

Change-Id: I96ce9ba8e7696d9e9ce89cc9004248e620df0654
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3367521
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#46230}