[cloudbuildhelper] Add `ignore` option to `copy` step.
It can be used to skip copying some files. Uses .gitignore
syntax (hence the name).

Also default ${inputsdir} to ${contextdir} if unset.

These changes are needed to add the following to the Monorail
manifest:

  - copy: "${inputsdir}/lib"
    dest: "${contextdir}/lib"
    ignore: ["*.pyc", "*.pyo"]

Monorail needs `lib`, but this directory is in .gitignore and
doesn't get copied to the tarball by default. But when we copy it
explicitly with `copy`, we still need to ignore *some* files there.
There's no existing .gitignore patterns for them, because the entire
directory is already in .gitignore.

Defaulting ${inputsdir} to ${contextdir} is needed to avoid confusing
(but still technically working) step like

  - copy: "${contextdir}/lib"
    dest: "${contextdir}/lib"

R=iannucci@chromium.org, tikhonov@google.com
BUG=b/204239298

Change-Id: I0763f984da3bcf6d7348315cb913de35d8772578
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3382355
Reviewed-by: Oleg Tikhonov <tikhonov@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#46305}