Add support for specifying multiple commits to bootstrapper.
Accept an additional property $bootstrap/trigger that a bootstrapped
builder can set to ensure that a triggered builder will read its
properties from the same revisions if it is also bootstrapped. This is
necessary because the build's gitiles commit may be for a repo that is
not involved in the bootstrapping process if a builder's config is
located in one repo while being triggered on commits to another repo.

Bug: 1210566, 1285293
Change-Id: I642e020c08ed4681f9892b9ace9af1da2ad0144c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3381551
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#46314}