[Weetbix] Auto-exonerate invocation-blocking failures in non-failing invocations.
Put simply, if the build did not fail, the test can't be responsible
for the build (or its enclosing CV run, if applicable) failing.
Therefore, regardless of whether an accurate set of exonerations
was uploaded to ResultDB, the impact of invocation-blocking failures
can be taken as mitigated if the build not fail.
(By the recipe, in case of a passed build, another infra failure,
in case of build with infra failure, or by virtue of the cancellation
in case of build cancellation).

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: I44587d2649ae0e3027359f227d773e1250ed297c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3388256
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#46379}