[crosskylabadmin] Add pool detection for tasks to CSA.
When enabling PARIS for labstations, we need the ability to control
whether a labstation is eligible for the PARIS flow by using pool
("label-pool") on the labstation's UFS entry. This CL adds the ability
to add an allowlist of pools to the CSA config.

BUG=b:206044795
TEST=make test

Change-Id: I40f01c72518f4c90967e302ccfca8d43c88fd07c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3381875
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#46429}