phosphorus: convert dut topology to host info
As part of topology rollout, this is the first attempt to convert dut
topology to host info. This is guarded under a flag which is only turned
on in dev env. There will be more iterations to complete this effort
based on findings from running tests.

BUG=b:201424819

Change-Id: I9df676d959d5bd4031ddeff889ea3a631456c04b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3399407
Reviewed-by: Jared Loucks <jaredloucks@google.com>
Commit-Queue: Azizur Rahman <azrahman@google.com>
Cr-Commit-Position: refs/heads/main@{#46444}