[crosadm] Use the default PRPC options in the UFS client.
Prior to this CL, the default PRPC options, including the user agent,
were not being passed to the UFS client. This means that we always
report a version of 1.0.0 instead of the true version shown in the
site.go file. See b:215410393 for details.

BUG=b:215410393
TEST=make test

Change-Id: If9ca3c778b7c40f196cea67a1dd5dcbd3460961f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3404553
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#46487}