[crosadm] Add a field to control on a per-task basis the UFS error policy.
Integrating UFS and CrOS Skylab Admin is an ongoing effort, some bugs
such as b:215550098 and b:215410393 track integration issues related to
using the bot ID instead of the DUT ID and passing version information.
This CL adds configuration settings to the CrOSSkylabAdmin config so
that we can control on a per-type-of-task basis how strict we should be
regarding UFS errors.

BUG=b:215410393
TEST=make test

Change-Id: I5e5febc6bc4f4f50b1909e9d5b61b3b139424c27
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3404991
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#46492}