recovery: Add toggle servod's "servo_pd_role" Exec func:
This cl add the exec function for toggle the servod's "servo_pd_role"
cmd from "snk" to "src" number of times. The number of times it needs to
toggle is based on the parameter passed in from the "config-servo" plan.

This cl includes the servo repair action for toggle the "servo_pd_role"
cmd in servod only once. And it also added this servo repair action to
the corresponding "recovery" field inside the "config-servo" plan.

BUG=b:198819047
TEST=manual test

Change-Id: I10594e09f8c77bd57bcca46537e4f7445ebef0f2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3374078
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Yunzhi Yu <yunzhiyu@google.com>
Cr-Commit-Position: refs/heads/main@{#46509}