[crosadm] Use percentage even when all labstations are opted in.
The current implementation of CrOSSkylabAdmin ignores the probability
completely when we've elected to opt in all labstations. This is a
problem because the slow rollout will actually be very fast.

Add logic to consult our random value and the threshold on the path
where we've opted all the labstations in.

BUG=b:206044795
TEST=make test

Change-Id: I82ea1868400c4278adeb94fe99fa5eb3cfe6c7d1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3405113
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#46514}