[skylab] Fix valid pool name heuristic
This heuristic is used in
`cmd/skylab/internal/userinput/device_specs.go`, which is used by shivas
and possibly other tools.

The actual regex in question includes a typo `A-z` instead of `A-Z`. Fix
this. This leads to, for example, the string "a[" being classified as a
valid pool name when it is not a valid pool name.

BUG=None
TEST=make test

Change-Id: Ifea10870a0664bb2862bff9587c313fa3e59f20f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3408006
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#46557}