[recovery]: Verify for warm reset pin only for servo V3 and micro types.
This commit is a part of the PARIS project.

This commit adds execs that checks for servo types V3 and micro. Then,
it alters the action to check for warm reset pin to use these two
execs.

The gist is that we need to check for warm reset pins only for the
servo types V3 and micro. This CL alters the action for warm reset pin
about this.

Bug=b:198809511
TEST=build, manual test, unittest

Change-Id: If463ce9a4bb6f08f6028753fa1943af78793555c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3413092
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#46570}