[crosadm] Interpret the rollout config value as a rollout permille.
As described in the bug, CrOSSklyabAdmin currently interprets the
rollout permille incorrectly as a target score to beat, rather than as a
target score to be lower than. This causes the rollout ratio to be
unexpectedly "flipped". See bug for details.

This CL adds a random-with-consistent-seed test to guard against this
kind of flip in the future.

BUG=b:216499840
TEST=make test

Change-Id: I39f91ba81a3c4e48a0ffa24ab74d215dd6321f14
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3418136
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#46634}