[Weetbix] Add schema for only triggering reclustering on predicate change.
Currently, any change to the rule triggers re-clustering. To avoid this,
we need to decouple the versioning of the parts of the rule relevant
to re-clustering (the rule predicate, consisting of RuleDefinition and
IsActive) and the versioning of the rule for audit/UI purposes.

This change sets up the necessary schema and begins to populate it.
A follow up change will start using the new timestamps to drive
re-clustering.

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: Id0837b70e6284fb9de5b2cc92965d2f3c6cbed99
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3426864
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#46835}