recovery: Remove "servo_host_servod_restart" dependencies.
This cl removes the dependencies for the servo host servod restart
action in the config file as well as in the servo_execs file.

All the existed servo_host_servod_restart action is replaced with the
servod stop action instead. Since, inside the recovery action, after the
servod is stopped, the plan will be run again from the top including the
servod start action which is the same as action "servo_host_servod_restart".

BUG=b:198819047
TEST=manual test

Change-Id: I6a72855b42b47e8d07d496f86e41fbe0c1e0cd83
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3440566
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Yunzhi Yu <yunzhiyu@google.com>
Cr-Commit-Position: refs/heads/main@{#46868}