recovery: Make rebootExec in cros_execs.go to reboot only in DUT's hostname instead of resource name.
This cl changes the rebootExec func to make the runner inside runs only
on the DUT.Name as the resourceName.

In the case of the config_labstation.go, the args.DUT.Name is the
labstation hostname so this will not impact the execution of the
labstation plan.

BUG=b:198819047
TEST=manual test

Change-Id: I16b92e72d5619793c7cddecd92bc236f58a0cbef
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3445290
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Yunzhi Yu <yunzhiyu@google.com>
Cr-Commit-Position: refs/heads/main@{#46915}