[drone-agent] Add logic to abbreviate and truncate path names.
The temporary directory path is currently too long, which is causing
Paris to fail for DUTs with hostnames longer than 37 characters.

To get around this, we will abbreviate hostnames that use the
word1-word2-word3-word4-word5 ... naming scheme by keeping all numbers
at the end and the first character of the beginning of the word.

BUG=b:218349208
TEST=make test

Change-Id: I4ef6e44e4acd4f26bfc93814ef4ea68c0fed1319
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3445291
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#46927}