[Weetbix] Improve bug priority update descriptions.
Provide reasons for bug priority changes.

Also fixes an issue with the application of bug priority hysteresis
whereby criteria of lower priorities (which needed to be satisfied
in order to reach a higher priority) were not evaluated in
determining if the bug priority needed to change.

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: I47292775b2736896c9a8c52767e1d4eb2eb297c8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3443079
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#46977}