[paris] Check for nil metrics instance when writing tasks.
Check for nil instance of metrics client when writing tasks.

In order to avoid crashing recovery, metrics methods should return an
error instead of panicking when the method receiver is nil. I also
updated the doc comments for each function to clarify that checking for
nil is the still the caller's responsibility. Calling these methods on a
nil object happens to be harmless, but is explicitly ruled out as a
guaranteed behavior.

BUG=b:217746479
TEST=make test

Change-Id: Ia0af49afbe3d427f8a443a236793599b654d74b6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3440568
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#46982}