[infra] Utility for collecting ml features for RTS
This is a cli for collecting the test stability and recreating file
graph distances to create a csv for generating a ml model for test
selection. There is a lot of room for improvement and parallelism but
getting a good set of starting data to work on training is more
important right now.


Bug: 1289329
Change-Id: I2633815865f95c562514025a7593879467fb931b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3405008
Reviewed-by: Gregory Guterman <guterman@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#47005}