[crosadm] Add reason field to stable version RPC.
As part of b:219072576, we are adding satlab-specific logic to the
stable version RPC. In the future, the reason field will be populated
with comments about the satlab code path as well. This is the main
reason for having it, for providing incontrovertible proof that the
satlab code path was taken, as well as information about whether we used
a generic board+model entry or used a hostname-specific override.

BUG=b:219072576
TEST=make test

Change-Id: I7b6ec8b545b44e98b66603954edfaca1c3ba1203
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3461757
Reviewed-by: Prasad Vuppalapu <prasadv@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#47076}