recovery: Fix servoCheckServodControlExec func when just logging the servod command value.
Currently, this servoCheckServodControlExec will only print the value
for the servod command if the value is a string, but for other type, it will print the empty string value.

This cl will change it into log the other types of value along with
their type to the output.

BUG=b:220018485
TEST=manual test

Change-Id: I39d82fec99cfcb99e385412489b36c0c20b16555
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3469390
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Yunzhi Yu <yunzhiyu@google.com>
Cr-Commit-Position: refs/heads/main@{#47157}