[Weetbix] Enable ingestion for multiple projects.
Enable ingestion for all configured Weetbix projects except for
chromium milestone projects.

Switch detection of builds that are part of presubmit runs
to user_agent:cq tag. Verify that the behaviour is the same for
chromium builds by logging warnings if there is a difference
and reverting to the old behaviour.

BUG=1243174
TEST=INTEGRATION_TESTS=1 go test infra/appengine/weetbix/...

Change-Id: I128dc007f3e018d581b793145a709d292a82f26a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3473613
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#47186}