recovery: Fix of the ssh proxy failure after device reboot.
Since the connection between local environment to the lab is flacky
especially after the device reboot itself, the originally created ssh
proxy connection will not work afterwards.

In order to fix this issue, this cl introduces the middle proxy
connection layer which is a DUT in the lab. This way we make sure the
DUT is always alive during the device reboot process, then, the proxy
ssh connection will work after the device reboot.

BUG=b:219763331
TEST=manual test

Change-Id: Icdac1f682348756d5925693e25c990940ae071f8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3466433
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Yunzhi Yu <yunzhiyu@google.com>
Cr-Commit-Position: refs/heads/main@{#47196}