[Weetbix] Add analysis of cluster impact before Weetbix (and FindIt) exoneration.
Where decisions are made to exonerate test failures, it should not
be based on data that includes those very decisions, to avoid
negative feedback loops. (Tests are exonerated because there are
impactful unexonerated failures, then after a while, because of that
very exoneration, there are none, so it ceases to be exonerated...
until the impact manifests again.)

Similiarly, while FindIt is making decisions to exonerate test faiures,
Weetbix should not be excluding this impact when determining whether
to file bugs, to avoid FindIt bugs interfering with Weetbix bug filing.

BUG=1243174

Change-Id: Ic444c73096c7c6271d4a99a79d020ef76b90eb39
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3479049
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#47261}