recovery: Fix isLastProvisionSuccessfulExec func.
In the last code refactor, it accidentally change the condition for
returning errors in this Exec func.

This Exec func will return error only when the command "test -f" cannot
find the provision failed file. Therefore, for a good DUT, this
provision failed file will not existed in the disk. Then, the condition
must be if the command cannot find this provision fail file, then this
verifier passed.

BUG=b:221091083
TEST=manual test

Change-Id: I46bf5dbaff0ade8e6607bcb55ba221a6cb0d85d4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3485285
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Yunzhi Yu <yunzhiyu@google.com>
Cr-Commit-Position: refs/heads/main@{#47286}