[crosadm] [satlab] Add explicit test case for looking up host information by board and model given a hostname.
CrOSSkylabAdmin has undergone a few changes since the test fakes were
defined. The current inventory system cannot be faked with the builtin
tests currently.

To support this test, create an override variable for the helper
function getDUTInfo. This is a better approach in the short/medium-term
than updating the test fakes and test helpers because many of them refer
to and are based on the git-based inventory system, which has not
existed for nearly two years.

BUG=b:219072576
TEST=make test

Change-Id: I92e2c22972b2447aa21ac374a5e2c66fccb1a2f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3487360
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#47401}