Remove tandrii@chromium.org from recipes/recipes/recipe_simulation.expected/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ibd8b1f11d638a73e30ff442d0d9e1c267b5a0b9e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3500868
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Cr-Commit-Position: refs/heads/main@{#47453}