Remove shapiroc@chromium.org from go/src/infra/tools/dirmd/proto/chromeos/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I74d92b881613ba24f7af9d67ba11c747742ad54e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3503884
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: George Engelbrecht <engeg@google.com>
Commit-Queue: George Engelbrecht <engeg@google.com>
Cr-Commit-Position: refs/heads/main@{#47455}