Remove mikenichols@chromium.org from go/src/infra/cros/cmd/prototype-tlw/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ic4177aa383c9b880c560648b79091ac2a97cae6e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3499319
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: George Engelbrecht <engeg@google.com>
Commit-Queue: George Engelbrecht <engeg@google.com>
Cr-Commit-Position: refs/heads/main@{#47459}