Remove mmoss@chromium.org from go/src/infra/appengine/statsui/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Iecc0d6b951d9110f021e48c17d15034312f9ea6f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3499318
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#47461}