Remove mmoss@chromium.org from OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I067322f7c6a472ad3d0e2db86145b207cad3f896
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3503360
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Sean McCullough <seanmccullough@google.com>
Commit-Queue: Sean McCullough <seanmccullough@google.com>
Cr-Commit-Position: refs/heads/main@{#47463}