Remove tandrii@chromium.org from recipes/recipes/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I2e34b266ff7ec8a8041d9a3e7771f716c5cd337e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3499317
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#47467}