Remove shapiroc@chromium.org from go/src/infra/appengine/crosskylabadmin/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I361c5b3c80e17d21bfb9e7633cdc3954a069e084
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3503106
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#47472}