Remove mikenichols@chromium.org from go/src/infra/qscheduler/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ifc3eee0a9dca98da150de78a7613698cff61fce5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3503924
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: Andrew Lamb <andrewlamb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#47495}