Remove mikenichols@chromium.org from go/src/infra/libs/skylab/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ic8866c597d12cf8a7c8e93dc4ad9abe7ae8665ff
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3504894
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: Andrew Lamb <andrewlamb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#47496}