Add Git Watcher correctly to the MD blocklist
Git Watcher email addresses were added to the markdown blocklist, but
this didn't work because in the case of Git Watcher, the author text is
the name instead of the email address.

This fixes the formatting of Git Watcher messages when markdown is
enabled, since Git Watcher messages aren't written in markdown.

It also leaves the email addresses in case the behavior to show the name
changes in the future.

Change-Id: I9b32a57f1b321bf98dea017e71641a5b9b2be8c7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3387335
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/main@{#47547}