recovery: Fix of the config_servo file: Expected value of the "servo_pd_role".
Currently, in the critical action: "servod_servo_pd", the expected value
is "snk", but according to the python verifier code, it should be "src"
instead. The python logic indicates that if it is "snk", then it will
raise the error. Therefore, the epected value needs to be the reverse of
the "snk" -> "src".

BUG=b:218718971
TEST=manual test

Change-Id: I9f0ca035a209cbdbf52d280548ba367f192ed633
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3508609
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Yunzhi Yu <yunzhiyu@google.com>
Cr-Commit-Position: refs/heads/main@{#47552}