recovery: Add feature for servod communication includes timeout options.
Currently, the servod communication has a default timeout of 10s, which
is not enough for some of the servod command to execute and return. For
example, "dut-control -p xxxx image_usbkey_dev" will have a runtime of
nearly 1 min.

This cl introduces the timeout for servod communication. When creating
servod using info.NewServod(), the default timeout will be the same as
the action timeout specified in the config file.

BUG=b:222358559
TEST=manual test

Change-Id: Ia80aeef805c7ce5b0b3b6c685e187d2b67847cf2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3501423
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Yunzhi Yu <yunzhiyu@google.com>
Cr-Commit-Position: refs/heads/main@{#47553}