compilator_watcher sets status to CANCELED when context is canceled
If the context is canceled, processErr() will set the build status to
CANCELED and return nil instead of the actual context canceled err.
This is because luciexe will automatically set the build status
to FAILURE if an err is returned. Returning nil will keep the build
status as canceled.
See runUserCode() in
https://source.chromium.org/chromium/infra/infra/+/main:go/src/go.chromium.org/luci/luciexe/exe/exe.go

It will also check for any context errors before fetching the compilator
build in copySteps(). If it sees a context err, it'll break out of the
for loop and return the context err.

Bug: 1304475
Change-Id: Ibeea0ffece47686619206a7183cd4e33e9087181
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3513884
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Stephanie Kim <kimstephanie@google.com>
Cr-Commit-Position: refs/heads/main@{#47580}