[labpack] Treat permission denied errors as non-transient when uploading
The labpack recipe (which was arguably misconfigured) looped endlessly
when attempting to upload to a google storage bucket for which it didn't
have permission. See b:223230183 for details.

This CL changes the implementation of upload to take into account
permission denied errors as a non-transient error. Non-transient errors
block retries.

BUG=b:223230183
TEST=make test

Change-Id: I6234751def1f47d391c67ec1da4d73f7f6b67e5b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3515572
Reviewed-by: Jared Loucks <jaredloucks@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#47611}