[karte] Use lowercase identifiers in doc comments for unexported identifiers.
As mentioned in this CL
(https://chromium-review.googlesource.com/c/infra/infra/+/3514574),
change doc comments for unexported identifiers to use the identifier
itself (i.e. don't change "foo" to "Foo" just because it's at the
beginning of a sentence).

There is precedent for doing this in the Go standard library, see for
example:
https://cs.opensource.google/go/x/tools/+/refs/tags/v0.1.9:cmd/godoc/main.go
.

BUG=None
TEST=make test

Change-Id: I887594a9adbb452182c371aac2c639ee84206b4f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3523441
Reviewed-by: Ashish Gandhi <ashishgandhi@google.com>
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#47638}