[cr-buildbucket] _compute_bbagent: generate constant length command for real builds only
if computing the command for a task_def for `led get-builder`, the build doesn't have a real id. So _compute_bbagent should not generate a constant length command in this case.

This is a short term fix to unblock reenabling BBAgent GetBuild mode.

After this change and https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3518406 are landed,
we could reenable BBAgent GetBuild mode.

But to fully remove the other mode, there are 2 remaining blockers:
- `led get-builder` stop using swarmbucket.get_task_def
- led creates real buildbucket build

Bug: 1219018
Change-Id: I9d74a3de2b92766c69eb537ba1e0b03b6acfb4ac
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3517621
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#47659}