[Weetbix] Add proper auto-bug filing kill switch.
The current approach of stopping auto-bug filing by changing
the monorail hostname is unintuitive and has other side-effects,
like causing the cron job to fail (triggering alerts), and
causing monorail RPCs calls to be made against the Weetbix
instance, in turn triggering metrics to be collected for non-existent
RPCs.

Adding a flag will make it obvious to a trooper how to disable this
part of the service in case of a problem (e.g. bug-filing runaway).

Also fix up global metrics reporting for this part of Weetbix.

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: I1fb0ccf3d37ccfac2127c4a15dd6167846fab4a5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3534980
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#47756}