gerrit: Clean up gerrit tests
This CL adds client_fake.go. The previous method for testing in
dependent packages involved mocking out the underlying RPC calls
for the gerrit package, which involved knowing things about the
impl of this package. This is undesirable both because it prevents
us from treating the gerrit package as an API and because it is not
generalizable -- if a method is added to the package that cannot
be mocked via the mock gitiles RPC client, we have no way of mocking.

Note that we still do want to test the implementation of methods in
client.go -- that's why client_test.go does the full request/response
mocking. It's just not appropriate to do this in external files.

BUG=b:224575508
TEST=go test

Change-Id: I2d23dbc0ee0b64a0ed4ba0f8d641f1d35c891cdf
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3541112
Reviewed-by: Benjamin Shai <bshai@google.com>
Reviewed-by: Madeleine Hardt <hardtmad@google.com>
Commit-Queue: Jack Neus <jackneus@google.com>
Cr-Commit-Position: refs/heads/main@{#47833}