[karte] Add example-create-update-action to CLI
Add a simple command that creates and then immediately updates an action
in Karte.

This command talks to the dev Karte instance. It is part of the
investigation of the bug: 215608994.

Running this command shows us exactly how an update in Karte modifies
the original record.

BUG=b:215608994
TEST=./karte example-create-update-action

Change-Id: I81af858411d5404e9572d17728b6cee6bee35164
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3541116
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#47853}